Standardize platform-specific code pattern
Refactor resourcecheck to match filecheck pattern:
- Common interface + struct in resource_common.go (no build tag)
- Platform-specific methods in resource_unix.go and resource_windows.go
- Document Pattern 6 in CLAUDE.md
