+++ pkg/resourcecheck/resource_windows.go
@@ -2,41 +2,14 @@
 
 package resourcecheck
 
-import (
-	"fmt"
-	"runtime"
-)
+import "errors"
 
-// ResourceChecker abstracts system resource detection for testability.
-type ResourceChecker interface {
-	// FreeDiskSpace returns free disk space in bytes at the given path.
-	FreeDiskSpace(path string) (uint64, error)
-
-	// AvailableMemory returns available memory in bytes.
-	// In containers, this respects cgroup limits.
-	AvailableMemory() (uint64, error)
-
-	// NumCPUs returns the number of available CPUs.
-	NumCPUs() int
-}
-
-// RealResourceChecker implements ResourceChecker using actual system calls.
-type RealResourceChecker struct{}
-
-// FreeDiskSpace returns free disk space in bytes.
-// Not implemented on Windows.
+// FreeDiskSpace is not supported on Windows.
 func (r *RealResourceChecker) FreeDiskSpace(path string) (uint64, error) {
-	return 0, fmt.Errorf("disk space check not supported on Windows")
+	return 0, errors.New("disk space check not supported on Windows")
 }
 
-// AvailableMemory returns available memory in bytes.
-// Not implemented on Windows.
+// AvailableMemory is not supported on Windows.
 func (r *RealResourceChecker) AvailableMemory() (uint64, error) {
-	return 0, fmt.Errorf("memory check not supported on Windows")
-}
-
-// NumCPUs returns the number of available CPUs.
-// Go's runtime.NumCPU() already respects container CPU limits.
-func (r *RealResourceChecker) NumCPUs() int {
-	return runtime.NumCPU()
+	return 0, errors.New("memory check not supported on Windows")
 }
